refactor: remove vision routing for native Composer image input
Drop the text-only image description path now that Composer 2.5
supports images natively, and clean related config, docs, and tests.
