+++ tests/config.test.ts
@@ -15,27 +15,22 @@ import {
   migrateLegacyConfig,
   saveConfig,
 } from '../src/config.js';
-import { useTempHome } from './vision/helpers.js';
+import { useTempHome } from './stateTestHelpers.js';
 
 const setupHome = useTempHome();
 const FEATURE_CONFIG = {
   ...DEFAULT_CONFIG,
   imagine: { enabled: false },
-  vision: { ...DEFAULT_CONFIG.vision, maxImages: 2 },
 };
 
 function paths(home: string) {
   return {
-    cache: join(home, '.pi', 'grok-cli', 'vision-cache.json'),
     config: join(home, '.pi', 'grok-cli', 'config.json'),
     data: join(home, '.pi', 'grok-cli'),
-    debug: join(home, '.pi', 'grok-cli-vision-debug.log'),
     imagine: join(home, '.pi', 'grok-cli-imagine.json'),
-    legacyCache: join(home, '.pi', 'grok-cli-vision-cache.json'),
     pi: join(home, '.pi'),
     quota: join(home, '.pi', 'grok-cli-quota.json'),
     tools: join(home, '.pi', 'grok-cli-tools.json'),
-    vision: join(home, '.pi', 'grok-cli-vision.json'),
   };
 }
 
@@ -80,18 +75,13 @@ describe('Grok CLI configuration', () => {
     const home = setupHome();
     writeConsolidatedConfig(home, {
       version: 2,
-      vision: { enabled: 'yes', maxImages: -1, cacheEnabled: false },
+      imagine: { enabled: 'yes' },
     });
 
     const loaded = loadConfig();
 
     expect(loaded.config.imagine).toEqual(DEFAULT_CONFIG.imagine);
-    expect(loaded.config.vision).toEqual({
-      ...DEFAULT_CONFIG.vision,
-      cacheEnabled: false,
-    });
-    expect(loaded.warning).toMatch(/enabled must be true or false/);
-    expect(loaded.warning).toMatch(/maxImages must be a positive integer/);
+    expect(loaded.warning).toMatch(/imagine\.enabled must be true or false/);
   });
 
   it('falls back to legacy settings without overwriting an unsupported version', () => {
@@ -108,39 +98,21 @@ describe('Grok CLI configuration', () => {
     expect(existsSync(paths(home).imagine)).toBe(true);
   });
 
-  it('migrates both current legacy files and leaves non-config data untouched', () => {
+  it('migrates the current Imagine legacy file and leaves non-config data untouched', () => {
     const home = setupHome();
     mkdirSync(paths(home).pi, { recursive: true });
     writeJson(paths(home).imagine, { enabled: false });
-    writeJson(paths(home).vision, {
-      enabled: false,
-      model: 'grok-build',
-      maxImages: 2,
-      cacheEnabled: false,
-      cacheMaxEntries: 25,
-    });
-    for (const path of [paths(home).legacyCache, paths(home).tools, paths(home).quota]) {
+    for (const path of [paths(home).tools, paths(home).quota]) {
       writeJson(path, { keep: true });
     }
-    writeFileSync(paths(home).debug, 'keep');
 
     expect(migrateLegacyConfig()).toEqual({});
     expect(loadConfig().config).toEqual({
       ...DEFAULT_CONFIG,
       imagine: { enabled: false },
-      vision: {
-        enabled: false,
-        model: 'grok-build',
-        maxImages: 2,
-        cacheEnabled: false,
-        cacheMaxEntries: 25,
-      },
     });
     expect(existsSync(paths(home).imagine)).toBe(false);
-    expect(existsSync(paths(home).vision)).toBe(false);
-    expect(existsSync(paths(home).legacyCache)).toBe(false);
-    expect(JSON.parse(readFileSync(paths(home).cache, 'utf8'))).toEqual({ keep: true });
-    for (const path of [paths(home).tools, paths(home).quota, paths(home).debug]) {
+    for (const path of [paths(home).tools, paths(home).quota]) {
       expect(existsSync(path)).toBe(true);
     }
   });
@@ -223,38 +195,29 @@ describe('Grok CLI configuration', () => {
     expect(existsSync(paths(home).imagine)).toBe(false);
   });
 
-  it.each([
-    ['Imagine', 'imagine'],
-    ['vision', 'vision'],
-  ] as const)('migrates a lone %s legacy file and defaults the other section', (_label, kind) => {
+  it('migrates a lone Imagine legacy file', () => {
     const home = setupHome();
     mkdirSync(paths(home).pi, { recursive: true });
-    if (kind === 'imagine') writeJson(paths(home).imagine, { enabled: false });
-    if (kind === 'vision') writeJson(paths(home).vision, { maxImages: 2 });
+    writeJson(paths(home).imagine, { enabled: false });
 
     expect(migrateLegacyConfig()).toEqual({});
     expect(loadConfig().config).toEqual({
       ...DEFAULT_CONFIG,
-      imagine: kind === 'imagine' ? { enabled: false } : DEFAULT_CONFIG.imagine,
-      vision:
-        kind === 'vision' ? { ...DEFAULT_CONFIG.vision, maxImages: 2 } : DEFAULT_CONFIG.vision,
+      imagine: { enabled: false },
     });
   });
 
-  it('preserves all legacy files and skips migration when one is malformed', () => {
+  it('preserves a malformed legacy file and skips migration', () => {
     const home = setupHome();
     mkdirSync(paths(home).pi, { recursive: true });
-    writeJson(paths(home).imagine, { enabled: false });
-    writeFileSync(paths(home).vision, '{ nope');
+    writeFileSync(paths(home).imagine, '{ nope');
 
     const migration = migrateLegacyConfig();
 
     expect(migration.warning).toMatch(/Could not read/);
     expect(existsSync(paths(home).config)).toBe(false);
     expect(existsSync(paths(home).imagine)).toBe(true);
-    expect(existsSync(paths(home).vision)).toBe(true);
-    expect(loadConfig().config.imagine.enabled).toBe(false);
-    expect(loadConfig().config.vision).toEqual(DEFAULT_CONFIG.vision);
+    expect(loadConfig().config.imagine).toEqual(DEFAULT_CONFIG.imagine);
   });
 
   it('preserves legacy files and removes temporary output when migration cannot write', () => {
@@ -277,13 +240,10 @@ describe('Grok CLI configuration', () => {
       imagine: { enabled: false },
     });
     writeJson(paths(home).imagine, { enabled: true });
-    writeJson(paths(home).vision, { maxImages: 2 });
 
     expect(migrateLegacyConfig()).toEqual({});
     expect(loadConfig().config.imagine.enabled).toBe(false);
-    expect(loadConfig().config.vision.maxImages).toBe(DEFAULT_CONFIG.vision.maxImages);
     expect(existsSync(paths(home).imagine)).toBe(false);
-    expect(existsSync(paths(home).vision)).toBe(false);
     expect(migrateLegacyConfig()).toEqual({});
   });
