+++ tests/storage.test.ts
@@ -6,12 +6,9 @@ import {
   getConfigPath,
   getGrokCliDirectory,
   getLegacyConfigPath,
-  getLegacyVisionCachePath,
   getQuotaCachePath,
-  getVisionCachePath,
 } from '../src/storage.js';
-import { getCachePath } from '../src/vision/cache.js';
-import { useTempHome } from './vision/helpers.js';
+import { useTempHome } from './stateTestHelpers.js';
 
 const setupHome = useTempHome();
 
@@ -25,64 +22,46 @@ describe('Grok CLI storage', () => {
 
     expect(getGrokCliDirectory()).toBe(join(home, '.pi', 'grok-cli'));
     expect(getConfigPath()).toBe(join(home, '.pi', 'grok-cli', 'config.json'));
-    expect(getVisionCachePath()).toBe(join(home, '.pi', 'grok-cli', 'vision-cache.json'));
     expect(getQuotaCachePath()).toBe(join(home, '.pi', 'grok-cli', 'quota-cache.json'));
     expect(loadConfig()).toEqual({ config: DEFAULT_CONFIG });
     expect(existsSync(getGrokCliDirectory())).toBe(false);
   });
 
-  it('migrates the consolidated config and vision cache after verified writes', () => {
+  it('migrates the consolidated config after a verified write', () => {
     const home = setupHome();
     mkdirSync(join(home, '.pi'), { recursive: true });
     writeJson(getLegacyConfigPath(), { ...DEFAULT_CONFIG, imagine: { enabled: false } });
-    writeJson(getLegacyVisionCachePath(), {
-      version: 1,
-      entries: { cached: { description: 'keep' } },
-    });
 
     expect(migrateLegacyConfig()).toEqual({});
 
     expect(loadConfig().config.imagine.enabled).toBe(false);
     expect(JSON.parse(readFileSync(getConfigPath(), 'utf8')).version).toBe(2);
-    expect(JSON.parse(readFileSync(getVisionCachePath(), 'utf8'))).toEqual({
-      version: 1,
-      entries: { cached: { description: 'keep' } },
-    });
     expect(existsSync(getLegacyConfigPath())).toBe(false);
-    expect(existsSync(getLegacyVisionCachePath())).toBe(false);
-    expect(getCachePath()).toBe(getVisionCachePath());
   });
 
-  it('keeps new files authoritative and preserves conflicting legacy files', () => {
+  it('keeps the new config authoritative and preserves a conflicting legacy file', () => {
     setupHome();
     mkdirSync(getGrokCliDirectory(), { recursive: true });
     writeJson(getConfigPath(), DEFAULT_CONFIG);
     writeJson(getLegacyConfigPath(), { ...DEFAULT_CONFIG, imagine: { enabled: false } });
-    writeJson(getVisionCachePath(), { version: 1, entries: { current: {} } });
-    writeJson(getLegacyVisionCachePath(), { version: 1, entries: { legacy: {} } });
 
     const migration = migrateLegacyConfig();
 
     expect(migration.warning).toContain(getLegacyConfigPath());
-    expect(migration.warning).toContain(getLegacyVisionCachePath());
     expect(loadConfig().config).toEqual(DEFAULT_CONFIG);
     expect(existsSync(getLegacyConfigPath())).toBe(true);
-    expect(existsSync(getLegacyVisionCachePath())).toBe(true);
   });
 
   it('falls back to legacy files when the destination directory cannot be created', () => {
     const home = setupHome();
     mkdirSync(join(home, '.pi'), { recursive: true });
     writeJson(getLegacyConfigPath(), { ...DEFAULT_CONFIG, imagine: { enabled: false } });
-    writeJson(getLegacyVisionCachePath(), { version: 1, entries: {} });
     writeFileSync(getGrokCliDirectory(), 'not a directory');
 
     const migration = migrateLegacyConfig();
 
     expect(migration.warning).toMatch(/Could not migrate/);
     expect(loadConfig().config.imagine.enabled).toBe(false);
-    expect(getCachePath()).toBe(getLegacyVisionCachePath());
     expect(existsSync(getLegacyConfigPath())).toBe(true);
-    expect(existsSync(getLegacyVisionCachePath())).toBe(true);
   });
 });
